renamed packages and resuse predefinded functional interfaces
