--- Java/src/main/java/com/evolvetechnology/timecost_calculation/TimeCostCalculator.java
@@ -1,8 +1,0 @@
-package com.evolvetechnology.timecost_calculation;
-
-import java.time.LocalTime;
-
-@FunctionalInterface
-public interface TimeCostCalculator {
-  int getCostFor(LocalTime localTime);
-}
