+++ Java/src/main/java/com/evolvetechnology/timecostcalculators/CostInterval.java
@@ -1,4 +1,4 @@
-package com.evolvetechnology.timecost_calculation;
+package com.evolvetechnology.timecostcalculators;
 
 import java.time.LocalTime;
